Add permanently delete user by id

+++ OjtPortal/Repositories/UserRepo.cs
@@ -83,6 +83,7 @@ namespace OjtPortal.Repositories
         public async Task<User?> DeleteByIdAsync(int id)
         {
             var (user, _) = await GetUserByIdAsync(id);
+            if (user == null) return null;
             _context.Users.Remove(user!);
             await _context.SaveChangesAsync();
             return user;
